Allow ZMQ_FLAGS to be picked up from environment
In order to cross compile erlzmq2 for other plaforms, the libzmq
configure script needs to be invoked with --host=xxx.  The makefile
for the nif passed the contents of ZMA_FLAGS to configure, however
it was overriding it.  By using += we can append to it and inherit
an initial value from the environment.

To cross compile erlzmq2 for say ARM, you first set ZMQ_FLAGS as
follows before invoking `rebar compile`:

`ZMQ_FLAGS=--host=arm-none-linux-gnueabi`